// techmedicine-frontend-angular13/src/app/dashboard/patients/models/patient.ts
export interface Patient {
  id: number;
  name: string;
  surname: string;
  birthDate: string;
  gender: string;
  rg: string;
  cpf: string;
  homePhone: string;
  mobilePhone: string;
  email: string;
  cep: string;
  city: string;
  state: string;
  address: string;
  number: string;
  district: string;
  complement: string;
}